09-serverless/pyproject.toml
[project]
name = "09-serverless"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.5",
    "onnx>=1.20.0",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchstudio>=0.0.2",
    "torchvision>=0.24.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
